requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asgi-lifespan==2.1.0
asyncpg==0.28.0
certifi==2023.5.7
click==8.1.5
dictdiffer==0.9.0
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
packaging==23.1
pluggy==1.2.0
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
pydantic==1.10.11
pydantic_core==2.3.0
pypika-tortoise==0.1.6
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.11.8
tortoise-orm==0.19.3
typing_extensions==4.7.1
uvicorn==0.22.0
